+++ src/simtools/runners/simtools_runner.py
@@ -7,7 +7,7 @@ from pathlib import Path
 
 import simtools.utils.general as gen
 from simtools import dependencies
-from simtools.data_model.metadata_collector import MetadataCollector
+from simtools.data_model import workflow_metadata
 from simtools.io import ascii_handler
 from simtools.job_execution import job_manager
 
@@ -63,9 +63,6 @@ def run_applications(args_dict, logger):
                 app_activity_id = app_configuration.get("activity_id") or gen.uuid()
                 app_configuration["activity_id"] = app_activity_id
                 associated_activities.append({"name": app, "activity_id": app_activity_id})
-                metadata_file = _get_model_parameter_metadata_file(app, app_configuration)
-                if metadata_file is not None:
-                    model_parameter_metadata_files.append(metadata_file)
 
                 logger.info(f"Running application: {app}")
                 result = job_manager.submit(
@@ -75,6 +72,9 @@ def run_applications(args_dict, logger):
                     configuration=app_configuration,
                     runtime_environment=run_time,
                 )
+                metadata_file = _get_model_parameter_metadata_file(app, app_configuration)
+                if metadata_file is not None:
+                    model_parameter_metadata_files.append(metadata_file)
                 file.write("=" * 80 + "\n")
                 file.write(
                     f"Application: {app}\nSTDOUT:\n{result.stdout}\nSTDERR:\n{result.stderr}\n"
@@ -82,21 +82,23 @@ def run_applications(args_dict, logger):
         finally:
             workflow_end = datetime.now(UTC)
             workflow_end = max(workflow_end, workflow_start)
-            workflow_metadata = _build_workflow_metadata(
-                args_dict=args_dict,
-                workflow_activity_id=workflow_activity_id,
-                workflow_start=workflow_start,
-                workflow_end=workflow_end,
-                runtime_environment=runtime_environment_snapshot,
-                workflow_site=workflow_site,
-                workflow_instrument=workflow_instrument,
-            )
-            _update_model_parameter_metadata_files(
-                model_parameter_metadata_files=model_parameter_metadata_files,
-                workflow_metadata=workflow_metadata,
-                associated_activities=associated_activities,
-                logger=logger,
-            )
+            if model_parameter_metadata_files:
+                workflow_activity = workflow_metadata.build_workflow_activity_metadata(
+                    args_dict=args_dict,
+                    workflow_activity_id=workflow_activity_id,
+                    workflow_start=workflow_start,
+                    workflow_end=workflow_end,
+                    runtime_environment=runtime_environment_snapshot,
+                    workflow_site=workflow_site,
+                    workflow_instrument=workflow_instrument,
+                )
+                for metadata_file in model_parameter_metadata_files:
+                    workflow_metadata.update_model_parameter_metadata_file(
+                        metadata_file=metadata_file,
+                        workflow_activity=workflow_activity,
+                        associated_activities=associated_activities,
+                        logger=logger,
+                    )
 
 
 def _read_application_configuration(configuration_file, steps, logger, workflow_activity_id=None):
@@ -162,29 +164,6 @@ def _read_application_configuration(configuration_file, steps, logger, workflow_
     )
 
 
-def _build_workflow_metadata(
-    args_dict,
-    workflow_activity_id,
-    workflow_start,
-    workflow_end,
-    runtime_environment,
-    workflow_site,
-    workflow_instrument,
-):
-    """Build workflow-level metadata dictionary with authoritative lifecycle timestamps."""
-    metadata_args = dict(args_dict)
-    metadata_args["label"] = "setting_workflow"
-    metadata_args["activity_id"] = workflow_activity_id
-    metadata_args["activity_start"] = workflow_start.isoformat(timespec="seconds")
-    metadata_args["activity_end"] = workflow_end.isoformat(timespec="seconds")
-    metadata_args["runtime_environment"] = deepcopy(runtime_environment)
-    metadata_args["site"] = workflow_site
-    metadata_args["instrument"] = workflow_instrument
-
-    collector = MetadataCollector(metadata_args, clean_meta=False)
-    return collector.get_top_level_metadata().get("cta", {})
-
-
 def _get_model_parameter_metadata_file(application, app_configuration):
     """Return expected metadata file for model-parameter submission applications."""
     if application != "simtools-submit-model-parameter-from-external":
@@ -197,51 +176,6 @@ def _get_model_parameter_metadata_file(application, app_configuration):
         return None
 
     return Path(output_path) / parameter / f"{parameter}-{parameter_version}.meta.yml"
-
-
-def _update_model_parameter_metadata_files(
-    model_parameter_metadata_files,
-    workflow_metadata,
-    associated_activities,
-    logger,
-):
-    """Inject workflow metadata into model-parameter metadata files."""
-    workflow_activity = deepcopy(workflow_metadata.get("activity", {}))
-
-    for metadata_file in model_parameter_metadata_files:
-        metadata_path = Path(metadata_file)
-        if not metadata_path.exists():
-            logger.debug(f"Model-parameter metadata file does not exist: {metadata_path}")
-            continue
-
-        metadata = ascii_handler.collect_data_from_file(metadata_path)
-        metadata = gen.change_dict_keys_case(metadata, True)
-        cta_meta = metadata.get("cta", {})
-        cta_meta["activity"] = deepcopy(workflow_activity)
-
-        context = cta_meta.setdefault("context", {})
-        context_associated = context.get("associated_activities") or []
-        context["associated_activities"] = _merge_associated_activities(
-            context_associated,
-            associated_activities,
-        )
-
-        metadata["cta"] = cta_meta
-        ascii_handler.write_data_to_file(metadata, metadata_path)
-        logger.info(f"Updated workflow metadata in {metadata_path}")
-
-
-def _merge_associated_activities(existing_activities, new_activities):
-    """Merge associated activities preserving order and uniqueness."""
-    merged_activities = []
-    seen = set()
-    for activity in [*existing_activities, *new_activities]:
-        key = (activity.get("name"), activity.get("activity_id"))
-        if key in seen:
-            continue
-        seen.add(key)
-        merged_activities.append(activity)
-    return merged_activities
 
 
 def _get_workflow_configuration_value(configurations, key):
