+++ tests/unit_tests/runners/test_simtools_runner.py
@@ -5,7 +5,6 @@ from pathlib import Path
 from unittest import mock
 
 import pytest
-import yaml
 
 import simtools.utils.general as gen
 from simtools.job_execution.job_manager import JobExecutionError
@@ -212,13 +211,14 @@ def test_run_applications_runs_and_logs(monkeypatch, tmp_test_directory):
         "simtools.runners.simtools_runner._read_application_configuration",
         mock.Mock(return_value=(mock_configurations, None, log_file_path, "wf-activity-id")),
     )
-    workflow_build_mock = mock.Mock(return_value={"activity": {"id": "wf-activity-id"}})
+    workflow_build_mock = mock.Mock(return_value={"id": "wf-activity-id"})
     workflow_update_mock = mock.Mock()
     monkeypatch.setattr(
-        "simtools.runners.simtools_runner._build_workflow_metadata", workflow_build_mock
+        "simtools.runners.simtools_runner.workflow_metadata.build_workflow_activity_metadata",
+        workflow_build_mock,
     )
     monkeypatch.setattr(
-        "simtools.runners.simtools_runner._update_model_parameter_metadata_files",
+        "simtools.runners.simtools_runner.workflow_metadata.update_model_parameter_metadata_file",
         workflow_update_mock,
     )
 
@@ -255,15 +255,8 @@ def test_run_applications_runs_and_logs(monkeypatch, tmp_test_directory):
     mock_logger.info.assert_any_call("Skipping application: app2")
     mock_logger.info.assert_any_call("Running application: app3")
     version_string_mock.assert_called_once_with([], include_software_versions=False)
-    workflow_build_mock.assert_called_once()
-    workflow_update_mock.assert_called_once()
-
-    associated = workflow_update_mock.call_args.kwargs["associated_activities"]
-    assert associated == [
-        {"name": "app1", "activity_id": "cfg-id-1"},
-        {"name": "app3", "activity_id": "cfg-id-3"},
-    ]
-    assert workflow_update_mock.call_args.kwargs["model_parameter_metadata_files"] == []
+    workflow_build_mock.assert_not_called()
+    workflow_update_mock.assert_not_called()
 
 
 def test_run_applications_passes_workflow_instrument_context(monkeypatch, tmp_test_directory):
@@ -275,9 +268,12 @@ def test_run_applications_passes_workflow_instrument_context(monkeypatch, tmp_te
     }
     mock_configurations = [
         {
-            "application": "app1",
+            "application": "simtools-submit-model-parameter-from-external",
             "run_application": True,
             "configuration": {
+                "parameter": "pm_photoelectron_spectrum",
+                "parameter_version": "2.0.1",
+                "output_path": "output/test_workflow",
                 "site": "North",
                 "telescope": "LSTN-design",
                 "activity_id": "cfg-id-1",
@@ -298,13 +294,14 @@ def test_run_applications_passes_workflow_instrument_context(monkeypatch, tmp_te
         "simtools.job_execution.job_manager.submit",
         mock.Mock(return_value=mock.Mock(stdout="ok", stderr="")),
     )
-    workflow_build_mock = mock.Mock(return_value={"activity": {"id": "wf-activity-id"}})
+    workflow_build_mock = mock.Mock(return_value={"id": "wf-activity-id"})
     workflow_update_mock = mock.Mock()
     monkeypatch.setattr(
-        "simtools.runners.simtools_runner._build_workflow_metadata", workflow_build_mock
+        "simtools.runners.simtools_runner.workflow_metadata.build_workflow_activity_metadata",
+        workflow_build_mock,
     )
     monkeypatch.setattr(
-        "simtools.runners.simtools_runner._update_model_parameter_metadata_files",
+        "simtools.runners.simtools_runner.workflow_metadata.update_model_parameter_metadata_file",
         workflow_update_mock,
     )
 
@@ -312,6 +309,7 @@ def test_run_applications_passes_workflow_instrument_context(monkeypatch, tmp_te
 
     assert workflow_build_mock.call_args.kwargs["workflow_site"] == "North"
     assert workflow_build_mock.call_args.kwargs["workflow_instrument"] == "LSTN-design"
+    workflow_update_mock.assert_called_once()
 
 
 def test_run_applications_handles_job_execution_exception(monkeypatch, tmp_test_directory):
@@ -341,11 +339,12 @@ def test_run_applications_handles_job_execution_exception(monkeypatch, tmp_test_
 
     monkeypatch.setattr("simtools.job_execution.job_manager.submit", mock_submit_failure)
     monkeypatch.setattr(
-        "simtools.runners.simtools_runner._build_workflow_metadata",
-        mock.Mock(return_value={"activity": {"id": "wf-activity-id"}}),
+        "simtools.runners.simtools_runner.workflow_metadata.build_workflow_activity_metadata",
+        mock.Mock(return_value={"id": "wf-activity-id"}),
     )
     monkeypatch.setattr(
-        "simtools.runners.simtools_runner._update_model_parameter_metadata_files", mock.Mock()
+        "simtools.runners.simtools_runner.workflow_metadata.update_model_parameter_metadata_file",
+        mock.Mock(),
     )
 
     with pytest.raises(JobExecutionError):
@@ -494,11 +493,12 @@ def test_run_applications_with_runtime_environment_ignored(monkeypatch, tmp_test
 
     monkeypatch.setattr("simtools.job_execution.job_manager.submit", mock_submit)
     monkeypatch.setattr(
-        "simtools.runners.simtools_runner._build_workflow_metadata",
-        mock.Mock(return_value={"activity": {"id": "wf-activity-id"}}),
+        "simtools.runners.simtools_runner.workflow_metadata.build_workflow_activity_metadata",
+        mock.Mock(return_value={"id": "wf-activity-id"}),
     )
     monkeypatch.setattr(
-        "simtools.runners.simtools_runner._update_model_parameter_metadata_files", mock.Mock()
+        "simtools.runners.simtools_runner.workflow_metadata.update_model_parameter_metadata_file",
+        mock.Mock(),
     )
 
     simtools_runner.run_applications(mock_args_dict, mock_logger)
@@ -572,11 +572,12 @@ def test_run_applications_with_empty_configuration_list(monkeypatch, tmp_test_di
     mock_submit = mock.Mock()
     monkeypatch.setattr("simtools.job_execution.job_manager.submit", mock_submit)
     monkeypatch.setattr(
-        "simtools.runners.simtools_runner._build_workflow_metadata",
-        mock.Mock(return_value={"activity": {"id": "wf-activity-id"}}),
+        "simtools.runners.simtools_runner.workflow_metadata.build_workflow_activity_metadata",
+        mock.Mock(return_value={"id": "wf-activity-id"}),
     )
     monkeypatch.setattr(
-        "simtools.runners.simtools_runner._update_model_parameter_metadata_files", mock.Mock()
+        "simtools.runners.simtools_runner.workflow_metadata.update_model_parameter_metadata_file",
+        mock.Mock(),
     )
 
     simtools_runner.run_applications(mock_args_dict, mock_logger)
@@ -626,32 +627,6 @@ def test_pull_image_raises_if_pull_fails(monkeypatch):
         simtools_runner._pull_image("podman", image)
 
 
-def test_build_workflow_metadata_uses_uncleaned_metadata(monkeypatch):
-    mock_collector = mock.Mock()
-    mock_collector.get_top_level_metadata.return_value = {"cta": {"activity": {"id": "wf-id"}}}
-    metadata_collector_cls = mock.Mock(return_value=mock_collector)
-    monkeypatch.setattr(
-        "simtools.runners.simtools_runner.MetadataCollector", metadata_collector_cls
-    )
-
-    metadata = simtools_runner._build_workflow_metadata(
-        args_dict={"config_file": "dummy.yml"},
-        workflow_activity_id="wf-id",
-        workflow_start=mock.Mock(isoformat=mock.Mock(return_value="2026-01-01T00:00:00+00:00")),
-        workflow_end=mock.Mock(isoformat=mock.Mock(return_value="2026-01-01T00:00:01+00:00")),
-        runtime_environment={"image": "test-image"},
-        workflow_site="North",
-        workflow_instrument="LSTN-design",
-    )
-
-    metadata_collector_cls.assert_called_once()
-    assert metadata_collector_cls.call_args.args[0]["activity_id"] == "wf-id"
-    assert metadata_collector_cls.call_args.args[0]["site"] == "North"
-    assert metadata_collector_cls.call_args.args[0]["instrument"] == "LSTN-design"
-    assert metadata_collector_cls.call_args.kwargs["clean_meta"] is False
-    assert metadata == {"activity": {"id": "wf-id"}}
-
-
 def test_get_model_parameter_metadata_file():
     config = {
         "output_path": "output/test",
@@ -664,40 +639,6 @@ def test_get_model_parameter_metadata_file():
     assert metadata_file == Path(
         "output/test/pm_photoelectron_spectrum/pm_photoelectron_spectrum-2.0.1.meta.yml"
     )
-
-
-def test_update_model_parameter_metadata_files(tmp_test_directory):
-    metadata_file = tmp_test_directory / "pm.meta.yml"
-    metadata_dict = {
-        "cta": {
-            "product": {"id": "prod-id"},
-            "activity": {"id": "old-id"},
-            "context": {"associated_activities": [{"name": "old", "activity_id": "old-id"}]},
-        }
-    }
-
-    metadata_file.write_text(yaml.safe_dump(metadata_dict), encoding="utf-8")
-    workflow_metadata = {"activity": {"id": "workflow-id", "name": "setting_workflow"}}
-    associated_activities = [
-        {"name": "app1", "activity_id": "a1"},
-        {"name": "app2", "activity_id": "a2"},
-    ]
-
-    simtools_runner._update_model_parameter_metadata_files(
-        model_parameter_metadata_files=[metadata_file],
-        workflow_metadata=workflow_metadata,
-        associated_activities=associated_activities,
-        logger=mock.Mock(),
-    )
-
-    updated = yaml.safe_load(metadata_file.read_text(encoding="utf-8"))
-    assert updated["cta"]["product"]["id"] == "prod-id"
-    assert updated["cta"]["activity"]["id"] == "workflow-id"
-    assert updated["cta"]["context"]["associated_activities"] == [
-        {"name": "old", "activity_id": "old-id"},
-        {"name": "app1", "activity_id": "a1"},
-        {"name": "app2", "activity_id": "a2"},
-    ]
 
 
 def test_get_workflow_configuration_value():
